backend/pyproject.toml
[project]
name = "realtime-chat"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "SQLAlchemy>=2.0",
  "psycopg[binary]>=3.1",
  "redis>=5.0",
]

[tool.uvicorn]
factory = false
